+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Fixed issue with retrieving type root
 - Fixed `exports` field to use `*` notation
 - Fixed resolution of `piral-cli` plugins with pnpm (#514)
-- Added `noEmptyRender` prop to `ExtensionSlot`
+- Added `emptySkipsRender` prop to `ExtensionSlot`
 - Added `order` prop to `ExtensionSlot`
 
 ## 0.14.23 (May 1, 2022)